+++ src/features/club-create/container/index.ts
@@ -1,1 +1,1 @@
-export {};
+export { FormContainer } from './FormContainer';
